refactor: Move NetexFeedParameters into netex graph builder
