+++ src/main/java/org/opentripplanner/graph_builder/GraphBuilderDataSources.java
@@ -21,12 +21,12 @@ import org.opentripplanner.graph_builder.module.ned.parameter.DemExtractParamete
 import org.opentripplanner.graph_builder.module.ned.parameter.DemExtractParametersBuilder;
 import org.opentripplanner.graph_builder.module.osm.parameters.OsmExtractParameters;
 import org.opentripplanner.graph_builder.module.osm.parameters.OsmExtractParametersBuilder;
+import org.opentripplanner.netex.config.NetexFeedParameters;
 import org.opentripplanner.standalone.config.BuildConfig;
 import org.opentripplanner.standalone.config.CommandLineParameters;
 import org.opentripplanner.standalone.config.api.OtpBaseDirectory;
 import org.opentripplanner.standalone.config.feed.GtfsFeedParameters;
 import org.opentripplanner.standalone.config.feed.GtfsFeedParametersBuilder;
-import org.opentripplanner.standalone.config.feed.NetexFeedParameters;
 import org.opentripplanner.util.OtpAppException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
