+++ src/main/java/org/opentripplanner/standalone/config/feed/TransitFeedConfig.java
@@ -5,6 +5,7 @@ import static org.opentripplanner.standalone.config.framework.json.OtpVersion.V2
 
 import java.util.List;
 import org.opentripplanner.graph_builder.model.DataSourceConfig;
+import org.opentripplanner.netex.config.NetexFeedParameters;
 import org.opentripplanner.standalone.config.framework.json.NodeAdapter;
 
 public class TransitFeedConfig {
