+++ src/main/java/org/opentripplanner/standalone/config/feed/TransitFeeds.java
@@ -4,6 +4,7 @@ import java.net.URI;
 import java.util.List;
 import javax.annotation.Nonnull;
 import org.opentripplanner.graph_builder.model.DataSourceConfig;
+import org.opentripplanner.netex.config.NetexFeedParameters;
 
 public record TransitFeeds(
   @Nonnull List<GtfsFeedParameters> gtfsFeeds,
